カバレッジ履歴の更新と静的解析の改善
- coverage_history.jsonに新しいカバレッジデータを追加し、古いデータを削除
- 静的解析の品質確認手順を統一し、ビルド確認コマンドを修正
- AlarmDisplayStateおよびAlarmLogicのメソッドを静的メソッドに変更し、コードの可読性を向上
- DateTimeInputStateでのnullptrチェックを明示化し、エラーハンドリングを強化
- PartialInputLogicでの戻り値を簡略化し、可読性を向上

+++ src/SettingsDisplayViewImpl.cpp
@@ -18,7 +18,7 @@ void SettingsDisplayViewImpl::showSettingsList(const std::vector<std::string>& i
     size_t displayCount = (items.size() < maxItems) ? items.size() : maxItems;
     
     for (size_t i = 0; i < displayCount; ++i) {
-        const int y_pos = startY + i * lineHeight;
+        const int y_pos = startY + static_cast<int>(i) * lineHeight;
         
         // 背景をクリア（選択状態に関係なく）
         disp->fillRect(0, y_pos - ALARM_BACKGROUND_OFFSET, SCREEN_WIDTH, lineHeight, TFT_BLACK);
